Double click on tray icon does play/pause Fix some small problems Begin support of Internet radios (experimental, use option --enable-radios to test it)
git-svn-id: https://ario-player.svn.sourceforge.net/svnroot/ario-player/ario@15 a3c249ba-b837-0410-a99a-8c7cbf76920c

+++ src/ario-shell-coverselect.c
@@ -512,7 +512,7 @@ ario_shell_coverselect_save_cover (ArioShellCoverselect *ario_shell_coverselect)
         gint *indice;
         gchar *data;
         const gchar *local_file;
-        guint size;
+        gsize size;
         gboolean ret;
 
         switch (gtk_notebook_get_current_page (GTK_NOTEBOOK (ario_shell_coverselect->priv->notebook))) {
